Add a link to a photo

+++ src/Emergence.Service/PhotoService.cs
@@ -112,7 +112,11 @@ namespace Emergence.Service
         public async Task<bool> RemovePhotoAsync(int id, string userId)
         {
             var photo = await GetPhotoAsync(id);
-            await _blobService.RemovePhotoAsync(photo.BlobPath);
+            if (!string.IsNullOrEmpty(photo.BlobPath))
+            {
+                await _blobService.RemovePhotoAsync(photo.BlobPath);
+            }
+
             var result = await _photoRepository.RemoveAsync(photo.AsStore());
             return result;
         }
